Added PushGatewayFormatter to format data
