+++ talaiot/src/main/kotlin/com/cdsap/talaiot/provider/PublishersProvider.kt
@@ -64,7 +64,8 @@ class PublishersProvider(
                         this,
                         logger,
                         SimpleRequest(logger),
-                        executor
+                        executor,
+                        PushGatewayFormatter()
                     )
                 )
             }
