+++ talaiot/src/main/kotlin/com/cdsap/talaiot/publisher/HybridPublisher.kt
@@ -61,7 +61,8 @@ class HybridPublisher(
                     publisherConfiguration,
                     logTracker,
                     SimpleRequest(logTracker),
-                    executor
+                    executor,
+                    PushGatewayFormatter()
                 )
             }
