+++ talaiot/src/main/kotlin/com/cdsap/talaiot/publisher/PushGatewayPublisher.kt
@@ -25,39 +25,13 @@ class PushGatewayPublisher(
     /**
      * Executor to schedule a task in Background
      */
-    private val executor: Executor
+    private val executor: Executor,
+    /**
+     * Formatter to format build and task data formatted as Pushgateway requirement
+     */
+    private val formatter: PushGatewayFormatter
 ) : Publisher {
     private val TAG = "PushGatewayPublisher"
-
-    fun getBuildMetricsContent(report: ExecutionReport): String {
-        val buildTags =
-                report.flattenBuildEnv()
-                        .map { (k, v) -> "${k.formatTagPublisher().replace(".", "_")}=\"${v.formatTagPublisher()}\"" }
-                        .joinToString(separator = ",")
-
-        return "${pushGatewayPublisherConfiguration.buildJobName}{$buildTags} ${report.durationMs}"
-    }
-
-    fun getTaskMetricsContent(report: ExecutionReport): String {
-        var contentTaskMetrics = ""
-        val taskProperties = report.customProperties.taskProperties.map { (k, v) ->
-            "${k.formatTagPublisher().replace(
-                    ".",
-                    "_"
-            )}=\"${v.formatTagPublisher()}\""
-        }.joinToString(separator = ",")
-
-        val properties = if (taskProperties.isNotBlank()) ",$taskProperties" else ""
-
-        report.tasks?.forEach {
-
-            val taskFormatted = it.taskPath.formatTagPublisher().replace("-","_")
-            contentTaskMetrics += "$taskFormatted{state=\"${it.state}\"" +
-                    ",module=\"${it.module}\",rootNode=\"${it.rootNode}\" $properties} ${it.ms}\n"
-        }
-
-        return contentTaskMetrics
-    }
 
     override fun publish(report: ExecutionReport) {
 
@@ -82,11 +56,11 @@ class PushGatewayPublisher(
 
 
             if (pushGatewayPublisherConfiguration.publishTaskMetrics) {
-               contentTaskMetrics += getTaskMetricsContent(report)
+               contentTaskMetrics += formatter.getTaskMetricsContent(report)
             }
 
             if (pushGatewayPublisherConfiguration.publishBuildMetrics) {
-                contentBuildMetrics += getBuildMetricsContent(report)
+                contentBuildMetrics += formatter.getBuildMetricsContent(report, pushGatewayPublisherConfiguration.buildJobName)
             }
             executor.execute {
